+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+* Added `getActiveAudioDevice` in `DefaultDeviceController`
+
 ## [0.11.0] - 2020-10-08
 
 ### Changed
@@ -11,6 +15,7 @@
 - Fixed a crash when opening ObjC demo app on iPhone 5/5c
 
 ## [0.10.0] - 2020-09-10
+
 
 ### Removed
 * **Breaking** Removed audio permission check in `DefaultAudioVideoController` which is performed in `DefaultAudioClientController`. For developers who has their own `AudioClientController` implementation, please make sure to check audio permission in `start()`.